Added color transition when dragging Articles.
A custom CSS class was created containing the rules
that are relevant to the overlay: width, height,
background, and text.

In the javascript (category.js.coffee) two event handlers
were bound (over, and out) each of these then uses the
jQuery toggle class to toggle the custom CSS onto the
overlay.

Transition was achieved by adding a css3 transition rule to
the main overlay class which is then triggered by the addition
or removal of the custom CSS.