Add proper titles and refactor category layouts
Implement a helper that will add proper page titles to every page
to add a title simply call the title() helper.

Refactor the current system for category layouts. Previously a render
statement was used to call a seperate view file for each item. Now the
system will always render as normal. Within the view file is a call to a
partial named for the category placed in a folder for the action. So the
edit page for the feature section is in "categories/edit/_feature.html.erb"
Should the partial be missing the system will rescue and show a view contained
in the main view file. The main advantages over the previous system are
a better organization of view files and the ability to still apply global
or conditional changes to the category and article views. This system also works
for articles to a sports article may render differently than a feature article
a possible use for this could be a "Photos" category in which the articles
would have a slideshow instead of a text centric layout.